+++ src/package/themes/default/index.ts
@@ -2,6 +2,7 @@ import base from './base';
 import button from './components/button';
 import card from './components/card';
 import input from './components/input';
+import searchInput from './components/search-input';
 import passwordInput from './components/password-input';
 import otpInput from './components/otp-input';
 import colorPicker from './components/color-picker';
@@ -58,6 +59,7 @@ export default {
         button,
         card,
         input,
+        searchInput,
         passwordInput,
         otpInput,
         colorPicker,